+++ polly/NQueens/impl/integer/copyreduced/NQueens_integer.h
@@ -1,11 +1,11 @@
 #include "Util.h"
 typedef struct{
-	long long r;
-	long long c;
+	int64_t r;
+	int64_t c;
 } POS;
 void printf_POS(POS* _pos);
 POS* copy_POS(POS* _pos);
 void free_POS(POS* _pos);
-long long conflict(POS* p, long long row, long long col);
-long long run(POS** queens, long long queens_size, long long n, long long dim);
+int64_t conflict(POS* p, int64_t row, int64_t col);
+int64_t run(POS** queens, size_t queens_size, int64_t n, int64_t dim);
 int main(int argc, char** args);
