+++ polly/newTicTacToe/impl/original/naive/newTicTacToe_original.h
@@ -3,15 +3,15 @@
 #define CIRCLE 1
 #define CROSS 2
 #define GAME [0,1,2,3,4,5,6,7,8]
-typedef long long nat;
-typedef long long Square;
+typedef int64_t nat;
+typedef int64_t Square;
 typedef struct{
 	_DECL_1DARRAY_MEMBER(pieces);
-	long long move;
+	int64_t move;
 } Board;
 void printf_Board(Board* _board);
 Board* copy_Board(Board* _board);
 void free_Board(Board* _board);
 Board* EmptyBoard();
-long long countOf(_DECL_1DARRAY_PARAM(pieces), long long s);
+int64_t countOf(_DECL_1DARRAY_PARAM(pieces), int64_t s);
 int main(int argc, char** args);
